smi_api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautiful-date==2.2.0
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-auditlog==2.2.2
django-cors-headers==3.13.0
django-prometheus==2.2.0
django-redis==5.3.0
django-threadlocals==0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-yasg==1.21.4
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.8.1
frozenlist==1.4.0
gcsa==2.0.1
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msgraph-core==0.2.2
multidict==6.0.4
mypy-extensions==0.4.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.16.0
protobuf==4.21.12
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.1
PyJWT==2.6.0
pymsteams==0.2.2
pyocclient==0.6
pyparsing==3.0.9
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.4.1
rsa==4.9
simple-salesforce==1.12.3
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
XlsxWriter==3.0.3
yarl==1.9.2